+++ examples/data-sources/iosxe_flow_monitor/data-source.tf
@@ -1,0 +1,3 @@
+data "iosxe_flow_monitor" "example" {
+  name = "MON1"
+}