+++ examples/resources/iosxe_flow_monitor/import.sh
@@ -1,0 +1,1 @@
+terraform import iosxe_flow_monitor.example "Cisco-IOS-XE-native:native/Cisco-IOS-XE-flow:flow/monitor=MON1"